requirements.txt
flask
requests
pymupdf
pytesseract
Pillow
opencv-python
transformers
torch
python-dotenv
